consistency
git-svn-id: https://svn.code.sf.net/p/web-erp/code/trunk@4635 09516a42-71f9-43be-811f-9b821cf0d246

+++ Locations.php
@@ -30,22 +30,20 @@ if (isset($_POST['submit'])) {
 	if ($_POST['CashSaleCustomer']!=''){
 
 		if ($_POST['CashSaleBranch']==''){
-                  prnMsg(_('A cash sale customer and branch are necessary to fully setup the counter sales functionality'),'error');
-                  $InputError =1;
-                } else { //customer branch is set too ... check it ties up with a valid customer
-                	$sql = "SELECT * FROM custbranch
-		               WHERE debtorno='" . $_POST['CashSaleCustomer'] . "'
-               		       AND branchcode='" . $_POST['CashSaleBranch'] . "'";
-
-      		        // echo $sql;
-                     	$result = DB_query($sql,$db);
-                        if (DB_num_rows($result)==0){
-      			   $InputError = 1;
-      			   prnMsg(_('The cash sale customer for this location must be defined with both a valid customer code and a valid branch code for this customer'),'error');
-      		        }
-                }
+			prnMsg(_('A cash sale customer and branch are necessary to fully setup the counter sales functionality'),'error');
+			$InputError =1;
+		} else { //customer branch is set too ... check it ties up with a valid customer
+			$sql = "SELECT * FROM custbranch
+					WHERE debtorno='" . $_POST['CashSaleCustomer'] . "'
+					AND branchcode='" . $_POST['CashSaleBranch'] . "'";
+		
+			$result = DB_query($sql,$db);
+			if (DB_num_rows($result)==0){
+				$InputError = 1;
+				prnMsg(_('The cash sale customer for this location must be defined with both a valid customer code and a valid branch code for this customer'),'error');
+			}
+		}
 	} //end of checking the customer - branch code entered
-
 
 	if (isset($SelectedLocation) AND $InputError !=1) {
 
@@ -56,25 +54,24 @@ if (isset($_POST['submit'])) {
 			$_POST['Managed'] = 0;
 		}
 
-		$sql = "UPDATE locations SET
-				loccode='" . $_POST['LocCode'] . "',
-				locationname='" . $_POST['LocationName'] . "',
-				deladd1='" . $_POST['DelAdd1'] . "',
-				deladd2='" . $_POST['DelAdd2'] . "',
-				deladd3='" . $_POST['DelAdd3'] . "',
-				deladd4='" . $_POST['DelAdd4'] . "',
-				deladd5='" . $_POST['DelAdd5'] . "',
-				deladd6='" . $_POST['DelAdd6'] . "',
-				tel='" . $_POST['Tel'] . "',
-				fax='" . $_POST['Fax'] . "',
-				email='" . $_POST['Email'] . "',
-				contact='" . $_POST['Contact'] . "',
-				taxprovinceid = '" . $_POST['TaxProvince'] . "',
-				cashsalecustomer ='" . $_POST['CashSaleCustomer'] . "',
-                                cashsalebranch ='" . $_POST['CashSaleBranch'] . "',
-				managed = '" . $_POST['Managed'] . "'
-			WHERE loccode = '" . $SelectedLocation . "'";
-
+		$sql = "UPDATE locations SET loccode='" . $_POST['LocCode'] . "',
+									locationname='" . $_POST['LocationName'] . "',
+									deladd1='" . $_POST['DelAdd1'] . "',
+									deladd2='" . $_POST['DelAdd2'] . "',
+									deladd3='" . $_POST['DelAdd3'] . "',
+									deladd4='" . $_POST['DelAdd4'] . "',
+									deladd5='" . $_POST['DelAdd5'] . "',
+									deladd6='" . $_POST['DelAdd6'] . "',
+									tel='" . $_POST['Tel'] . "',
+									fax='" . $_POST['Fax'] . "',
+									email='" . $_POST['Email'] . "',
+									contact='" . $_POST['Contact'] . "',
+									taxprovinceid = '" . $_POST['TaxProvince'] . "',
+									cashsalecustomer ='" . $_POST['CashSaleCustomer'] . "',
+									cashsalebranch ='" . $_POST['CashSaleBranch'] . "',
+									managed = '" . $_POST['Managed'] . "'
+						WHERE loccode = '" . $SelectedLocation . "'";
+									
 		$ErrMsg = _('An error occurred updating the') . ' ' . $SelectedLocation . ' ' . _('location record because');
 		$DbgMsg = _('The SQL used to update the location record was');
 
@@ -111,42 +108,38 @@ if (isset($_POST['submit'])) {
 
 		/*SelectedLocation is null cos no item selected on first time round so must be adding a	record must be submitting new entries in the new Location form */
 
-		$sql = "INSERT INTO locations (
-					loccode,
-					locationname,
-					deladd1,
-					deladd2,
-					deladd3,
-					deladd4,
-					deladd5,
-					deladd6,
-					tel,
-					fax,
-					email,
-					contact,
-					taxprovinceid,
-					cashsalecustomer,
-					cashsalebranch,
-					managed
-					)
-			VALUES (
-				'" . $_POST['LocCode'] . "',
-				'" . $_POST['LocationName'] . "',
-				'" . $_POST['DelAdd1'] ."',
-				'" . $_POST['DelAdd2'] ."',
-				'" . $_POST['DelAdd3'] . "',
-				'" . $_POST['DelAdd4'] . "',
-				'" . $_POST['DelAdd5'] . "',
-				'" . $_POST['DelAdd6'] . "',
-				'" . $_POST['Tel'] . "',
-				'" . $_POST['Fax'] . "',
-				'" . $_POST['Email'] . "',
-				'" . $_POST['Contact'] . "',
-				'" . $_POST['TaxProvince'] . "',
-				'" . $_POST['CashSaleCustomer'] . "',
-        			'" . $_POST['CashSaleBranch'] . "',
-				'" . $_POST['Managed'] . "'
-			)";
+		$sql = "INSERT INTO locations (loccode,
+										locationname,
+										deladd1,
+										deladd2,
+										deladd3,
+										deladd4,
+										deladd5,
+										deladd6,
+										tel,
+										fax,
+										email,
+										contact,
+										taxprovinceid,
+										cashsalecustomer,
+										cashsalebranch,
+										managed )
+						VALUES ('" . $_POST['LocCode'] . "',
+								'" . $_POST['LocationName'] . "',
+								'" . $_POST['DelAdd1'] ."',
+								'" . $_POST['DelAdd2'] ."',
+								'" . $_POST['DelAdd3'] . "',
+								'" . $_POST['DelAdd4'] . "',
+								'" . $_POST['DelAdd5'] . "',
+								'" . $_POST['DelAdd6'] . "',
+								'" . $_POST['Tel'] . "',
+								'" . $_POST['Fax'] . "',
+								'" . $_POST['Email'] . "',
+								'" . $_POST['Contact'] . "',
+								'" . $_POST['TaxProvince'] . "',
+								'" . $_POST['CashSaleCustomer'] . "',
+				        		'" . $_POST['CashSaleBranch'] . "',
+								'" . $_POST['Managed'] . "')";
 
 		$ErrMsg =  _('An error occurred inserting the new location record because');
 		$DbgMsg =  _('The SQL used to insert the location record was');
@@ -252,7 +245,9 @@ if (isset($_POST['submit'])) {
 			echo '<br />' . _('There are') . ' ' . $myrow[0] . ' ' . _('stock movements with this Location code');
 
 		} else {
-			$sql= "SELECT COUNT(*) FROM locstock WHERE locstock.loccode='". $SelectedLocation . "' AND locstock.quantity !=0";
+			$sql= "SELECT COUNT(*) FROM locstock 
+					WHERE locstock.loccode='". $SelectedLocation . "' 
+					AND locstock.quantity !=0";
 			$result = DB_query($sql,$db);
 			$myrow = DB_fetch_row($result);
 			if ($myrow[0]>0) {
@@ -260,7 +255,8 @@ if (isset($_POST['submit'])) {
 				prnMsg(_('Cannot delete this location because location stock records exist that use this location and have a quantity on hand not equal to 0'),'warn');
 				echo '<br /> ' . _('There are') . ' ' . $myrow[0] . ' ' . _('stock items with stock on hand at this location code');
 			} else {
-				$sql= "SELECT COUNT(*) FROM www_users WHERE www_users.defaultlocation='" . $SelectedLocation . "'";
+				$sql= "SELECT COUNT(*) FROM www_users 
+						WHERE www_users.defaultlocation='" . $SelectedLocation . "'";
 				$result = DB_query($sql,$db);
 				$myrow = DB_fetch_row($result);
 				if ($myrow[0]>0) {
@@ -268,7 +264,8 @@ if (isset($_POST['submit'])) {
 					prnMsg(_('Cannot delete this location because it is the default location for a user') . '. ' . _('The user record must be modified first'),'warn');
 					echo '<br /> ' . _('There are') . ' ' . $myrow[0] . ' ' . _('users using this location as their default location');
 				} else {
-					$sql= "SELECT COUNT(*) FROM bom WHERE bom.loccode='" . $SelectedLocation . "'";
+					$sql= "SELECT COUNT(*) FROM bom 
+							WHERE bom.loccode='" . $SelectedLocation . "'";
 					$result = DB_query($sql,$db);
 					$myrow = DB_fetch_row($result);
 					if ($myrow[0]>0) {
@@ -276,7 +273,8 @@ if (isset($_POST['submit'])) {
 						prnMsg(_('Cannot delete this location because it is the default location for a bill of material') . '. ' . _('The bill of materials must be modified first'),'warn');
 						echo '<br /> ' . _('There are') . ' ' . $myrow[0] . ' ' . _('bom components using this location');
 					} else {
-						$sql= "SELECT COUNT(*) FROM workcentres WHERE workcentres.location='" . $SelectedLocation . "'";
+						$sql= "SELECT COUNT(*) FROM workcentres 
+								WHERE workcentres.location='" . $SelectedLocation . "'";
 						$result = DB_query($sql,$db);
 						$myrow = DB_fetch_row($result);
 						if ($myrow[0]>0) {
@@ -284,7 +282,8 @@ if (isset($_POST['submit'])) {
 							prnMsg( _('Cannot delete this location because it is used by some work centre records'),'warn');
 							echo '<br />' . _('There are') . ' ' . $myrow[0] . ' ' . _('works centres using this location');
 						} else {
-							$sql= "SELECT COUNT(*) FROM workorders WHERE workorders.loccode='" . $SelectedLocation . "'";
+							$sql= "SELECT COUNT(*) FROM workorders 
+									WHERE workorders.loccode='" . $SelectedLocation . "'";
 							$result = DB_query($sql,$db);
 							$myrow = DB_fetch_row($result);
 							if ($myrow[0]>0) {
@@ -292,7 +291,8 @@ if (isset($_POST['submit'])) {
 								prnMsg( _('Cannot delete this location because it is used by some work order records'),'warn');
 								echo '<br />' . _('There are') . ' ' . $myrow[0] . ' ' . _('work orders using this location');
 							}else {
-								$sql= "SELECT COUNT(*) FROM custbranch WHERE custbranch.defaultlocation='" . $SelectedLocation . "'";
+								$sql= "SELECT COUNT(*) FROM custbranch 
+										WHERE custbranch.defaultlocation='" . $SelectedLocation . "'";
 								$result = DB_query($sql,$db);
 								$myrow = DB_fetch_row($result);
 								if ($myrow[0]>0) {
@@ -310,13 +310,16 @@ if (isset($_POST['submit'])) {
 	if (! $CancelDelete) {
 
 		/* need to figure out if this location is the only one in the same tax province */
-		$result = DB_query("SELECT taxprovinceid FROM locations WHERE loccode='" . $SelectedLocation . "'",$db);
+		$result = DB_query("SELECT taxprovinceid FROM locations 
+							WHERE loccode='" . $SelectedLocation . "'",$db);
 		$TaxProvinceRow = DB_fetch_row($result);
-		$result = DB_query("SELECT COUNT(taxprovinceid) FROM locations WHERE taxprovinceid='" .$TaxProvinceRow[0] . "'",$db);
+		$result = DB_query("SELECT COUNT(taxprovinceid) FROM locations 
+							WHERE taxprovinceid='" .$TaxProvinceRow[0] . "'",$db);
 		$TaxProvinceCount = DB_fetch_row($result);
 		if ($TaxProvinceCount[0]==1){
 		/* if its the only location in this tax authority then delete the appropriate records in TaxAuthLevels */
-			$result = DB_query("DELETE FROM taxauthrates WHERE dispatchtaxprovince='" . $TaxProvinceRow[0] . "'",$db);
+			$result = DB_query("DELETE FROM taxauthrates 
+								WHERE dispatchtaxprovince='" . $TaxProvinceRow[0] . "'",$db);
 		}
 
 		$result= DB_query("DELETE FROM locstock WHERE loccode ='" . $SelectedLocation . "'",$db);
@@ -337,10 +340,11 @@ links to delete or edit each. These will call the same page again and allow upda
 or deletion of the records*/
 
 	$sql = "SELECT loccode,
-			locationname,
-			taxprovinces.taxprovincename as description,
-			managed
-		FROM locations INNER JOIN taxprovinces ON locations.taxprovinceid=taxprovinces.taxprovinceid";
+				locationname,
+				taxprovinces.taxprovincename as description,
+				managed
+			FROM locations INNER JOIN taxprovinces 
+			ON locations.taxprovinceid=taxprovinces.taxprovinceid";
 	$result = DB_query($sql,$db);
 
 	if (DB_num_rows($result)==0){
@@ -372,18 +376,18 @@ while ($myrow = DB_fetch_array($result)) {
 	}
 */
 	printf('<td>%s</td>
-		<td>%s</td>
-		<td>%s</td>
-		<td><a href="%sSelectedLocation=%s">' . _('Edit') . '</td>
-		<td><a href="%sSelectedLocation=%s&delete=1">' . _('Delete') . '</td>
-		</tr>',
-		$myrow['loccode'],
-		$myrow['locationname'],
-		$myrow['description'],
-		$_SERVER['PHP_SELF'] . '?',
-		$myrow['loccode'],
-		$_SERVER['PHP_SELF'] . '?',
-		$myrow['loccode']);
+			<td>%s</td>
+			<td>%s</td>
+			<td><a href="%sSelectedLocation=%s">' . _('Edit') . '</td>
+			<td><a href="%sSelectedLocation=%s&delete=1" onclick="return confirm(\'' . _('Are you sure you wish to delete this inventory location?') . '\');">' . _('Delete') . '</td>
+			</tr>',
+			$myrow['loccode'],
+			$myrow['locationname'],
+			$myrow['description'],
+			$_SERVER['PHP_SELF'] . '?',
+			$myrow['loccode'],
+			$_SERVER['PHP_SELF'] . '?',
+			$myrow['loccode']);
 
 	}
 	//END WHILE LIST LOOP
@@ -409,23 +413,23 @@ if (!isset($_GET['delete'])) {
 			_('Inventory') . '" alt="" />' . ' ' . $title . '</p>';
 
 		$sql = "SELECT loccode,
-				locationname,
-				deladd1,
-				deladd2,
-				deladd3,
-				deladd4,
-				deladd5,
-				deladd6,
-				contact,
-				fax,
-				tel,
-				email,
-				taxprovinceid,
-				cashsalecustomer,
-				cashsalebranch,
-				managed
-			FROM locations
-			WHERE loccode='" . $SelectedLocation . "'";
+					locationname,
+					deladd1,
+					deladd2,
+					deladd3,
+					deladd4,
+					deladd5,
+					deladd6,
+					contact,
+					fax,
+					tel,
+					email,
+					taxprovinceid,
+					cashsalecustomer,
+					cashsalebranch,
+					managed
+				FROM locations
+				WHERE loccode='" . $SelectedLocation . "'";
 
 		$result = DB_query($sql, $db);
 		$myrow = DB_fetch_array($result);
@@ -448,10 +452,12 @@ if (!isset($_GET['delete'])) {
 		$_POST['Managed'] = $myrow['managed'];
 
 
-		echo '<input type="hidden" name="SelectedLocation" value="' . $SelectedLocation . '">';
-		echo '<input type="hidden" name="LocCode" value="' . $_POST['LocCode'] . '">';
+		echo '<input type="hidden" name="SelectedLocation" value="' . $SelectedLocation . '" />';
+		echo '<input type="hidden" name="LocCode" value="' . $_POST['LocCode'] . '" />';
 		echo '<table class="selection">';
-		echo '<tr><th colspan=2><font size=3 color=blue>'._('Amend Location details').'</font></th></tr>';
+		echo '<tr>
+				<th colspan=2><font size=3 color=blue>'._('Amend Location details').'</font></th>
+			</tr>';
 		echo '<tr><td>' . _('Location Code') . ':</td><td>';
 		echo $_POST['LocCode'] . '</td></tr>';
 	} else { //end of if $SelectedLocation only do the else when a new record is being entered
@@ -511,27 +517,27 @@ if (!isset($_GET['delete'])) {
 	}
 
 	echo '<tr><td>' .  _('Location Name') . ':' . '</td>';
-	echo '<td><input type="Text" name="LocationName" value="'. $_POST['LocationName'] . '" size=51 maxlength=50></td></tr>';
+	echo '<td><input type="text" name="LocationName" value="'. $_POST['LocationName'] . '" size=51 maxlength="50" /></td></tr>';
 	echo '<tr><td>' . _('Contact for deliveries') . ':' . '</td>';
-	echo '<td><input type="Text" name="Contact" value="' . $_POST['Contact'] . '" size=31 maxlength=30></td></tr>';
+	echo '<td><input type="text" name="Contact" value="' . $_POST['Contact'] . '" size="31" maxlength="30" /></td></tr>';
 	echo '<tr><td>' .  _('Delivery Address 1') . ':' . '</td>';
-	echo '<td><input type="Text" name="DelAdd1" value="' . $_POST['DelAdd1'] . '" size=41 maxlength=40></td></tr>';
+	echo '<td><input type="text" name="DelAdd1" value="' . $_POST['DelAdd1'] . '" size="41" maxlength="40" /></td></tr>';
 	echo '<tr><td>' . _('Delivery Address 2') . ':' . '</td>';
-	echo '<td><input type="Text" name="DelAdd2" value="' .  $_POST['DelAdd2'] . '" size=41 maxlength=40></td></tr>';
+	echo '<td><input type="text" name="DelAdd2" value="' .  $_POST['DelAdd2'] . '" size="41" maxlength="40" /></td></tr>';
 	echo '<tr><td>' .  _('Delivery Address 3') . ':' . '</td>';
-	echo '<td><input type="Text" name="DelAdd3" value="' .  $_POST['DelAdd3'] . '" size=41 maxlength=40></td></tr>';
+	echo '<td><input type="text" name="DelAdd3" value="' .  $_POST['DelAdd3'] . '" size="41" maxlength="40" /></td></tr>';
 	echo '<tr><td>' .  _('Delivery Address 4') . ':' . '</td>';
-	echo '<td><input type="Text" name="DelAdd4" value="' . $_POST['DelAdd4'] . '" size=41 maxlength=40></td></tr>';
+	echo '<td><input type="text" name="DelAdd4" value="' . $_POST['DelAdd4'] . '" size="41" maxlength=40 /></td></tr>';
 	echo '<tr><td>' .  _('Delivery Address 5') . ':' . '</td>';
-	echo '<td><input type="Text" name="DelAdd5" value="' . $_POST['DelAdd5'] . '" size=21 maxlength=20></td></tr>';
+	echo '<td><input type="text" name="DelAdd5" value="' . $_POST['DelAdd5'] . '" size="21" maxlength="20" /></td></tr>';
 	echo '<tr><td>' . _('Delivery Address 6') . ':' . '</td>';
-	echo '<td><input type="Text" name="DelAdd6" value="' . $_POST['DelAdd6'] . '" size=16 maxlength=15></td></tr>';
+	echo '<td><input type="text" name="DelAdd6" value="' . $_POST['DelAdd6'] . '" size="16" maxlength="15" /></td></tr>';
 	echo '<tr><td>' .  _('Telephone No') . ':' . '</td>';
-	echo '<td><input type="Text" name="Tel" value="' . $_POST['Tel'] . '" size=31 maxlength=30></td></tr>';
+	echo '<td><input type="text" name="Tel" value="' . $_POST['Tel'] . '" size="31" maxlength="30" /></td></tr>';
 	echo '<tr><td>' .  _('Facsimile No') . ':' . '</td>';
-	echo '<td><input type="Text" name="Fax" value="' . $_POST['Fax'] . '" size=31 maxlength=30></td></tr>';
+	echo '<td><input type="text" name="Fax" value="' . $_POST['Fax'] . '" size="31" maxlength="30" /></td></tr>';
 	echo '<tr><td>' .  _('Email') . ':' . '</td>';
-	echo '<td><input type="Text" name="Email" value="' . $_POST['Email'] . '" size=31 maxlength=55></td></tr>';
+	echo '<td><input type="text" name="Email" value="' . $_POST['Email'] . '" size="31" maxlength="55" /></td></tr>';
 
 	echo '<td>' .  _('Tax Province') . ':' . '</td><td><select name="TaxProvince">';
 
@@ -546,11 +552,11 @@ if (!isset($_GET['delete'])) {
 
 	echo '</select></td></tr>';
 	echo '<tr><td>' . _('Default Counter Sales Customer Code') . ':' . '</td>';
-	echo '<td><input type="Text" name="CashSaleCustomer" value="' . $_POST['CashSaleCustomer'] .
-		'" size=11 maxlength=10></td></tr>';
+	echo '<td><input type="text" name="CashSaleCustomer" value="' . $_POST['CashSaleCustomer'] .
+		'" size="11" maxlength="10" /></td></tr>';
 	echo '<tr><td>' . _('Counter Sales Branch Code') . ':' . '</td>';
-	echo '<td><input type="Text" name="CashSaleBranch" value="' . $_POST['CashSaleBranch'] .
-		'" size=11 maxlength=10></td></tr>';
+	echo '<td><input type="text" name="CashSaleBranch" value="' . $_POST['CashSaleBranch'] .
+		'" size="11" maxlength="10" /></td></tr>';
 	/*
 	This functionality is not written yet ...
 	<tr><td><?php echo _('Enable Warehouse Management') . ':'; ?></td>
